Exercises 4.15 to 4.23

+++ part4/bloglist/controllers/blog.js
@@ -1,27 +1,59 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const newUser = await User.findById(user.id)
+
+  if (!newUser) {
+    return response.status(400).json({ error: 'userId invalid or missing' })
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: newUser._id
   })
 
-  await blog.save()
-  response.status(201).json(blog)
+  const savedBlog = await blog.save()
+  newUser.blogs = newUser.blogs.concat(savedBlog)
+  await newUser.save()
+
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
   const id = request.params.id
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'invalid token' })
+  }
+
+  const blog = await Blog.findById(id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (!(blog.user.toString() === user.id)) {
+    return response.status(401).json({ error: 'not authorized' })
+  }
 
   await Blog.findByIdAndDelete(id)
   response.status(204).end()
